pyproject.toml
[project]
name = "task-qianbo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface>=0.0.1",
    "huggingface-hub>=1.3.5",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "pycocotools>=2.0.11",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "trimesh>=4.11.1",
]
